Selectively downgrade missing sync peer to WARNING
This uses comparison of more recent and less recent polls to determine
whether reachability is improving.  If so, a missing sync peer is
considered WARNING rather than CRITICAL, on the assumption that ntpd will
correct the problem shortly.  Thanks to Paul Collins for the idea and
initial implementation of this.

Update the test suite to check for this, and add python3 compatibility
for print statements.